requirements.txt
# Python 3.12.2
jieba==0.42.1
pandas==2.2.3
tqdm==4.67.0
rank_bm25==0.2.2
torch==2.5.1
transformers==4.46.2
python-dotenv==1.0.1
langchain==0.3.7
langchain_openai==0.2.6
pdfplumber==0.11.4
PyPDF2==3.0.1
fitz==0.0.1.dev2
frontend==0.0.3
google-generativeai==0.8.3